+++ CarAppBE/CarApp.Business/Interfaces/IRepository.cs
@@ -3,7 +3,7 @@
     public interface  IRepository<T>
     {
         public Task<IEnumerable<T>> GetAll();
-        public Task<T> GetById(long id);
+        public Task<T> GetById(int id);
         public Task<T> Add(T entity);
         public Task<T> Update(T entity);
         public Task Delete(T entity);
